update emacs install scripts, add file manifest

+++ scripts/install-emacs.sh
@@ -5,7 +5,7 @@
 # Author: Dan Harms <danielrharms@gmail.com>
 # Created: Saturday, July 25, 2015
 # Version: 1.0
-# Modified Time-stamp: <2015-07-25 13:26:21 dharms>
+# Modified Time-stamp: <2015-07-31 11:56:08 dan.harms>
 # Modified by: Dan Harms
 # Keywords: configuration
 
@@ -14,6 +14,8 @@ verbose=
 emacs=$EDITOR
 user=$USER
 int=emacs.tar
+manifest=.bk_manifest
+backup=emacs_bk.tar
 
 if [ $# -gt 0 ] ; then
    verbose=v
@@ -22,13 +24,24 @@ fi
 
 echo "Tarring .emacs.d into $int..."
 $tar c"$verbose"f $int --exclude=*.elc .emacs.d
-cp -f $int ~
+mv -f $int ~
 
 echo "Untarring $int into $HOME..."
 
 pushd ~
-rm -rf .emacs.d
-$tar -C $HOME -x"$verbose"pf $int
+
+if [ -d .emacs.d ] ; then
+   rm -f $backup
+   # todo
+   rm -rf .emacs.d
+   mkdir .emacs.d
+   if [ -r $backup ] ; then
+      $tar -xvpf $backup
+      rm -f $backup
+   fi
+fi
+
+tar --overwrite -x"$verbose"pf $int
 rm -f $int
 
 emacscomp.sh .emacs.d
